+++ lib/bond.ex
@@ -23,29 +23,17 @@ defmodule Bond do
   """
   @type assertion_expression :: {atom(), Macro.metadata(), list()}
 
-  @typedoc """
-  Subset of `Macro.Env` struct that excludes fields that, according to the documentation, "are
-  private to Elixir's macro expansion mechanism".
-  """
-  @type env :: %{
-          optional(:__struct__) => module(),
-          context: Macro.Env.context(),
-          context_modules: Macro.Env.context_modules(),
-          file: Macro.Env.file(),
-          function: Macro.Env.name_arity() | nil,
-          line: Macro.Env.line(),
-          module: module()
-        }
-
   @doc false
   defmacro __using__(_opts) do
-    Bond.Contracts.init(__CALLER__.module)
+    Bond.Compiler.init(__CALLER__.module)
 
     quote do
-      import Kernel, except: [@: 1, def: 2, defp: 2]
+      import Kernel, except: [@: 1]
       import Bond
 
-      @before_compile Bond.Contracts
+      @on_definition Bond.Compiler
+      @before_compile Bond.Compiler
+      @after_compile Bond.Compiler
     end
   end
 
@@ -62,10 +50,10 @@ defmodule Bond do
   defmacro @{pre_or_post, meta, [expression]} when pre_or_post in [:pre, :post] do
     if Keyword.keyword?(expression) do
       for {label, expression} <- expression do
-        Bond.Contracts.register_assertion(pre_or_post, expression, label, __CALLER__, meta)
+        Bond.Compiler.register_assertion(pre_or_post, expression, label, __CALLER__, meta)
       end
     else
-      Bond.Contracts.register_assertion(pre_or_post, expression, nil, __CALLER__, meta)
+      Bond.Compiler.register_assertion(pre_or_post, expression, nil, __CALLER__, meta)
     end
 
     :ok
@@ -74,19 +62,19 @@ defmodule Bond do
   # This clause handles @pre or @post assertions that have a label preceding them.
   defmacro @{pre_or_post, meta, [label, {_, _, _} = expression]}
            when (pre_or_post in [:pre, :post] and is_atom(label)) or is_binary(label) do
-    Bond.Contracts.register_assertion(pre_or_post, expression, label, __CALLER__, meta)
+    Bond.Compiler.register_assertion(pre_or_post, expression, label, __CALLER__, meta)
     :ok
   end
 
   # This clause handles @pre or @post assertions that have a label following them.
   defmacro @{pre_or_post, meta, [{_, _, _} = expression, label]}
            when (pre_or_post in [:pre, :post] and is_atom(label)) or is_binary(label) do
-    Bond.Contracts.register_assertion(pre_or_post, expression, label, __CALLER__, meta)
+    Bond.Compiler.register_assertion(pre_or_post, expression, label, __CALLER__, meta)
     :ok
   end
 
   defmacro @{:doc, meta, [value]} do
-    Bond.Contracts.register_doc(__CALLER__, meta, value)
+    Bond.Compiler.register_doc(__CALLER__, meta, value)
     :ok
   end
 
@@ -95,20 +83,6 @@ defmodule Bond do
     quote do
       Kernel.@(unquote(attr))
     end
-  end
-
-  @doc """
-  Override `Kernel.def/2` to support wrapping with preconditions and postconditions.
-  """
-  defmacro def(definition, body) do
-    Bond.Contracts.define_function_with_contract(__CALLER__, definition, body, true)
-  end
-
-  @doc """
-  Override `Kernel.defp/2` to support wrapping with preconditions and postconditions.
-  """
-  defmacro defp(definition, body) do
-    Bond.Contracts.define_function_with_contract(__CALLER__, definition, body, false)
   end
 
   @doc """
@@ -136,12 +110,12 @@ defmodule Bond do
 
   defmacro check(keyword_list) when is_list(keyword_list) do
     for {label, {_, meta, _} = expression} <- keyword_list do
-      Bond.Contracts.check_assertion(expression, label, __CALLER__, meta)
+      Bond.Compiler.check_assertion(expression, label, __CALLER__, meta)
     end
   end
 
   defmacro check({_, meta, _} = expression) do
-    Bond.Contracts.check_assertion(expression, nil, __CALLER__, meta)
+    Bond.Compiler.check_assertion(expression, nil, __CALLER__, meta)
   end
 
   @doc """
@@ -153,11 +127,11 @@ defmodule Bond do
 
   @spec check(assertion_label(), assertion_expression()) :: as_boolean(any())
   defmacro check(label, {_, meta, _} = expression) when is_atom(label) or is_binary(label) do
-    Bond.Contracts.check_assertion(expression, label, __CALLER__, meta)
+    Bond.Compiler.check_assertion(expression, label, __CALLER__, meta)
   end
 
   @spec check(assertion_expression(), assertion_label()) :: as_boolean(any())
   defmacro check({_, meta, _} = expression, label) when is_atom(label) or is_binary(label) do
-    Bond.Contracts.check_assertion(expression, label, __CALLER__, meta)
+    Bond.Compiler.check_assertion(expression, label, __CALLER__, meta)
   end
 end